refactor(config): move log related config into own file
