+++ internal/config/config.go
@@ -31,42 +31,6 @@ type Config struct {
 	Warnings []string       `json:"warnings,omitempty"`
 }
 
-type LogConfig struct {
-	Level     string      `json:"level,omitempty"`
-	File      string      `json:"file,omitempty"`
-	Encoding  LogEncoding `json:"encoding,omitempty"`
-	GRPCLevel string      `json:"grpc_level,omitempty"`
-}
-
-// LogEncoding is either console or JSON
-type LogEncoding uint8
-
-func (e LogEncoding) String() string {
-	return logEncodingToString[e]
-}
-
-func (e LogEncoding) MarshalJSON() ([]byte, error) {
-	return json.Marshal(e.String())
-}
-
-const (
-	_ LogEncoding = iota
-	LogEncodingConsole
-	LogEncodingJSON
-)
-
-var (
-	logEncodingToString = map[LogEncoding]string{
-		LogEncodingConsole: "console",
-		LogEncodingJSON:    "json",
-	}
-
-	stringToLogEncoding = map[string]LogEncoding{
-		"console": LogEncodingConsole,
-		"json":    LogEncodingJSON,
-	}
-)
-
 type UIConfig struct {
 	Enabled bool `json:"enabled"`
 }
@@ -292,12 +256,6 @@ func Default() *Config {
 }
 
 const (
-	// Logging
-	logLevel     = "log.level"
-	logFile      = "log.file"
-	logEncoding  = "log.encoding"
-	logGRPCLevel = "log.grpc_level"
-
 	// UI
 	uiEnabled = "ui.enabled"
